tested transformers

+++ src/anomaly_detection/encoders/ipaddress_encoder.py
@@ -40,7 +40,7 @@ class IPAddressEncoder(BaseEstimator, TransformerMixin):
             else:
                 transformed.append([1, *octets])
 
-        return np.asarray(transformed, dtype=np.uint16)
+        return np.asarray(transformed, dtype=np.float64)
 
     def get_feature_names_out(self, input_features=None):
         if input_features is None:
